+++ crates/uv-build/src/main.rs
@@ -1,10 +1,32 @@
-use anyhow::{Context, Result, bail};
 use std::env;
 use std::io::Write;
 use std::path::PathBuf;
 
+use anyhow::{Context, Result, bail};
+use tracing_subscriber::filter::LevelFilter;
+use tracing_subscriber::layer::SubscriberExt;
+use tracing_subscriber::util::SubscriberInitExt;
+use tracing_subscriber::{EnvFilter, Layer};
+
+use uv_logging::UvFormat;
+
 /// Entrypoint for the `uv-build` Python package.
 fn main() -> Result<()> {
+    // Support configuring the log level with `RUST_LOG` (shows only the error level by default) and
+    // color.
+    //
+    // This configuration is a simplified version of the uv logging configuration. When using
+    // uv_build through uv proper, the uv logging configuration applies.
+    let filter = EnvFilter::builder()
+        .with_default_directive(LevelFilter::OFF.into())
+        .from_env()
+        .context("Invalid RUST_LOG directives")?;
+    let stderr_layer = tracing_subscriber::fmt::layer()
+        .event_format(UvFormat::default())
+        .with_writer(std::sync::Mutex::new(anstream::stderr()))
+        .with_filter(filter);
+    tracing_subscriber::registry().with(stderr_layer).init();
+
     // Handrolled to avoid the large clap dependency
     let mut args = env::args_os();
     // Skip the name of the binary
