+++ crates/uv-logging/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "uv-logging"
+version = "0.0.1"
+edition = { workspace = true }
+rust-version = { workspace = true }
+homepage = { workspace = true }
+documentation = { workspace = true }
+repository = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+
+[lib]
+doctest = false
+
+[lints]
+workspace = true
+
+[dependencies]
+jiff = { workspace = true }
+owo-colors = { workspace = true }
+tracing-subscriber = { workspace = true }
+tracing = { workspace = true }
