+++ crates/uv/src/logging.rs
@@ -1,25 +1,19 @@
-use std::fmt;
 use std::str::FromStr;
 
 use anyhow::Context;
-use jiff::Timestamp;
-use owo_colors::OwoColorize;
-use tracing::{Event, Subscriber};
 #[cfg(feature = "tracing-durations-export")]
 use tracing_durations_export::{
     DurationsLayer, DurationsLayerBuilder, DurationsLayerDropGuard, plot::PlotConfig,
 };
 use tracing_subscriber::filter::Directive;
-use tracing_subscriber::fmt::format::Writer;
-use tracing_subscriber::fmt::{FmtContext, FormatEvent, FormatFields};
 use tracing_subscriber::layer::SubscriberExt;
-use tracing_subscriber::registry::LookupSpan;
 use tracing_subscriber::util::SubscriberInitExt;
 use tracing_subscriber::{EnvFilter, Layer, Registry};
 use tracing_tree::HierarchicalLayer;
 use tracing_tree::time::Uptime;
 
 use uv_cli::ColorChoice;
+use uv_logging::UvFormat;
 use uv_static::EnvVars;
 
 #[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
@@ -29,81 +23,6 @@ pub(crate) enum Level {
     DebugUv,
     TraceUv,
     TraceAll,
-}
-
-struct UvFormat {
-    display_timestamp: bool,
-    display_level: bool,
-    show_spans: bool,
-}
-
-/// See <https://docs.rs/tracing-subscriber/0.3.18/src/tracing_subscriber/fmt/format/mod.rs.html#1026-1156>
-impl<S, N> FormatEvent<S, N> for UvFormat
-where
-    S: Subscriber + for<'a> LookupSpan<'a>,
-    N: for<'a> FormatFields<'a> + 'static,
-{
-    fn format_event(
-        &self,
-        ctx: &FmtContext<'_, S, N>,
-        mut writer: Writer<'_>,
-        event: &Event<'_>,
-    ) -> fmt::Result {
-        let meta = event.metadata();
-        let ansi = writer.has_ansi_escapes();
-
-        if self.display_timestamp {
-            if ansi {
-                write!(writer, "{} ", Timestamp::now().dimmed())?;
-            } else {
-                write!(writer, "{} ", Timestamp::now())?;
-            }
-        }
-
-        if self.display_level {
-            let level = meta.level();
-            // Same colors as tracing
-            if ansi {
-                match *level {
-                    tracing::Level::TRACE => write!(writer, "{} ", level.purple())?,
-                    tracing::Level::DEBUG => write!(writer, "{} ", level.blue())?,
-                    tracing::Level::INFO => write!(writer, "{} ", level.green())?,
-                    tracing::Level::WARN => write!(writer, "{} ", level.yellow())?,
-                    tracing::Level::ERROR => write!(writer, "{} ", level.red())?,
-                }
-            } else {
-                write!(writer, "{level} ")?;
-            }
-        }
-
-        if self.show_spans {
-            let span = event.parent();
-            let mut seen = false;
-
-            let span = span
-                .and_then(|id| ctx.span(id))
-                .or_else(|| ctx.lookup_current());
-
-            let scope = span.into_iter().flat_map(|span| span.scope().from_root());
-
-            for span in scope {
-                seen = true;
-                if ansi {
-                    write!(writer, "{}:", span.metadata().name().bold())?;
-                } else {
-                    write!(writer, "{}:", span.metadata().name())?;
-                }
-            }
-
-            if seen {
-                writer.write_char(' ')?;
-            }
-        }
-
-        ctx.field_format().format_fields(writer.by_ref(), event)?;
-
-        writeln!(writer)
-    }
 }
 
 /// Configure `tracing` based on the given [`Level`], taking into account the `RUST_LOG` environment
@@ -183,18 +102,11 @@ pub(crate) fn setup_logging(
             )
             .init();
     } else {
-        // Regardless of the tracing level, show messages without any adornment.
-        let format = UvFormat {
-            display_timestamp: false,
-            display_level: true,
-            show_spans: false,
-        };
-
         tracing_subscriber::registry()
             .with(durations_layer)
             .with(
                 tracing_subscriber::fmt::layer()
-                    .event_format(format)
+                    .event_format(UvFormat::default())
                     .with_writer(writer)
                     .with_ansi(ansi)
                     .with_filter(filter),
