+++ docs/concepts/build-backend.md
@@ -237,6 +237,13 @@ must either be under the module root or in the appropriate
 [data directory](../reference/settings.md#build-backend_data). Most packages store small data in the
 module root alongside the source code.
 
+!!! tip
+
+    When using the uv build backend through a frontend that is not uv, such as pip or
+    `pythom -m build`, debug logging can be enabled through environment variables with
+    `RUST_LOG=uv=debug` or `RUST_LOG=uv=verbose`. When used through uv, the uv build backend shares
+    the verbosity level of uv.
+
 ### Include and exclude syntax
 
 Includes are anchored, which means that `pyproject.toml` includes only `<root>/pyproject.toml` and
